Force Activity launch in multiwindow mode
If Activity wants to launch in a sepcific multiwindow mode, such as
freeform, pip, pinned or split screen, we keep its mode.
If Activity wants to launch in fullscreen mode or other modes will
processed to fullscreen mode, we change the window mode to freeform mode
forcibly.
We should add switch/property to determine what time to apply this
change focibly later.

Change-Id: I7efb21782d8cfbd7719c83600b09792c8e6cce43
Signed-off-by: utzcoz <utzcoz@outlook.com>