Support restore function of multi-window
Add method enterFreeformMode in IActivityManager and Activity to provide
function to enter freeform mode for Activity. It will help to implement
restore function of window. When window is in fullscreen mode, and user
clicks the maximize button, the window will exit fullscreen mode and
enter freefor mode. Why maximize button? Because we don't have a restore
icon image, so keep to use the maximize icon to implement function
firstly.

The TaskRecord uses mLastNonFullscreenBounds to store the lastest non
fullscreen bounds, but it uses the reference of Rect, not new instance,
which will cause wrong size problem, when the origin Rect changes its
value. Also we fix the bug of ActivityStack.resize method that will not
adjust size of tasks to match that stack, the freeform specific stack.

TODO: Change maxmize button icon to restore icon.

Change-Id: I491cdd76dda26ea9b0d268da8b95d660478a11ac
Signed-off-by: utzcoz <utzcoz@outlook.com>