Disable app transition animation
In Android 9.0, system uses remote animation controller to control the
app transition animation, including app opening animation. Now the
Launcher3 plays the role to control the app opening animation. For
opening animation, the system uses the TaskStack's bounds as the final
position and size of window, and plays animation to archive the goal.
But when system gets the TaskStack's bounds of freeform, the bounds is
the same as display size, and doesn't be updated to real size, what
causes rigid app opening animation. To avoid this problem, I just close
the app transition aniamtion in config. I know this solution is rigid,
but I think the current app transition effect is ok without animation.

Change-Id: I9cce5c009387d346c836a039b53590852038ba01
Signed-off-by: utzcoz <utzcoz@outlook.com>