Fix drag-resizing jump of freeform
When starting to drag freeform window, the DecorView will create
BackdropFrameRenderer to drop back drop frame for freeform window.
The BackdropFrameRenderer will create a render node called back drop
render node, and set its bounds to window frame bounds. But in the
FrameBuilder.cpp in hwui will calculate the distance between back drop
render node bounds with frame contents drawing bounds, and translate
window content with this distance to adapt back drop position. But there
exists a problem, the back drop bounds is set with window frame bounds
what are relative to display, but the content drawing bounds are
relative to window left/top point. So it will cause the problem freeform
window will jump when drag-resizing, and come back to the correct size
after drag-resizing. So we change the back drop bounds to be relative
to window left/top point too to fix this problem.

Test: start settings from Taskbar, and drag shadow of window to resize
      window, and the window's left and top position keep the origin
      location and doesn't jump when resizing.

Signed-off-by: utzcoz <utzcoz@outlook.com>