Support persist window bounds
1. Add methods in WMS to store/restore window bounds in
   SharedPreferences.
2. When resizing/moving window, save its value.
3. When launching window, use saved value as launch bounds.

This patch also disable the launch bounds in ActivityOptions, because
the starter doesn't know the new window bounds after resizing. The
Taskbar sets the launch bounds to the center position of screen
forcibly, and it will cause saved bounds can't work. Before finding a
solution to let Taskbar to manage freeform window bounds, we will
disable launch bounds of ActivityOptions.

Signed-off-by: utzcoz <utzcoz@outlook.com>