Don't throw exception for setShelfHeight checking
If we change default overview recents component, the Launcher3QuickStep
will crash because it doesn't have STATE_BAR permission and is not
recents app when it tryies to invoke setShelfHeight. We will change
default recents app, and keep Launcher3QuickStep, so we should disable
throwing exception, and just do nothing for method invokeing.

Signed-off-by: utzcoz <utzcoz@outlook.com>